Wire up edit issue most of the way
In order to add a description to the issue, it
made sense to edit the existing issue. So, in order to
edit the issue, I abstracted the issue view model creation
to a factory, so we can dynamically bind to the model.
Then I started adding the update method to the API and
then I stopped.