Add a settings dialog
Currently stores everything a bit insecurely in localStorage,
should probably just limit the password to in-memory, only, really.

That means the abstraction for settings will need to change. However,
we should be able to hold on to the username and password field?

Or maybe we just present a login screen to the user on start?  If
the settings say to use Jira?  And only require a display name
otherwise?