flesh out the settings more
Added the settings object to the database, hooked up
the authorise step to the settings - we want to know the
username and password, but not hold on to it insecurely.

We'll probably send the auth header to the server, which
passes that auth header on to the Jira service.  I
think that should work regardless of which login method I
land on - basic / oauth? right?

I should probably test that the login combo works prior
to re-directing.