start reading issues from Jira
It's still pretty raw, and certainly doesn't display
the jira markup very nicely, but it's a thing, right.

The top fifty issues can be shown from Jira.

Still need to include the assignee details and display them
onscreen, and possibly intgerpret the Jira Markup better.  But after
that, it's a small leap to transition their state and I'm... MVP!