Migrate the client to the aurelia-cli
The Aurelia CLI has much better bundling support and does a lot
of the work for you.  Supposedly.  It also requires a lot of
non-standard setup in order to get things like async / await
to work in browsers that don't support it natively.

However, doing that has removed one of the tasks from the list
almost automatically - font awesome and bootstrap fonts are now
copied over correctly and aren't version-coded.  We can copy
the client/scripts folder and that would be a deployment.