Make the column headers sticky
Now that the layout of the board is a bit more (and yet less)
dynamic, I am free to move the status column header out of
the column itself.  That means it's not even part of the page
that gets considered for scrolling, so I don't have to do
anything fancy in javascript - I just leave it to the browser.