Generic implementation of markdown converter
Handles the basic cases in a generic way, can probably be extended to
deal with many other cases, but probably not tables, and maybe not
links.  So lists, it can probably be extended to handle lists.  Might be
some edge cases hiding, like a closing markdown tag as the final
character of a string, etc.