IssueTypes retrieved on the API will have colours
Abstracted the colour retrieval into its own module, which means that
there was basically no reason for the local cache to exist, which is
great, because I really wanted to get rid of it - it was holding me back
from multiple jira tenancy.  This also means that the issue type
dropdown can now show the colour marker in Jira, too.