Implemented basic epic search for local
Doesn't save the epic, because there's no field to save it into.  Have
to decide if / how I want to diverge from Jira as far as Epic DB
structure goes (separate table / type of Issue).  I see no reason for it
not to be a type of issue, other than it making integration with Jira
mildly easier.