Fixed issues in Epics on the second page not showing
The Jira API is not respecting the maxResults param for the Epics
endpoint, so it was not getting any epics on the second page.  As luck
would have it, though, we already have all the information we need, so
we'll just use that instead.  It requires a small amount of magice, but
that's probably fine.