Fixed inline markup issue
We would return the whole rest of the line as though it were markup when
we used a regex-based markup - heading, link or image.  This would cause
extra 'end of line' text to show.  Most of the time we got away with it,
because there was no rest of line.

It was exacerbated because we did this twice, so would end up with 4x
the fun.