Fix up for Jira's lack of respect for REST
Because Jira doesn't return the updated object, I have to go back to
the server and get it myself.

Also, I worked around the circular reference which meant changes to my
code.  But to be fair, I probably would have had the same problem when I
started having sub-tasks on the board.