Show the Project avatar on the Nav Panel
Of course, this is once again a CDN image, so it requires us to be
logged in to Jira via OAuth somewhere in the browser for the cookies to
work.  I'd call it myself if it weren't so terrible.

I should probably also not teach the UI about the settings, and just add
a setting for the local project, too.  Of course that will come with
autocomplete goodness, or something.  i.e. remove the Id requirement
from the request.