Include the recent growth rate in the estimate
If there were more issues resolved than created in the last n days
(currently set to 28) then we're probably approaching the end of the
work.  However, if we're still finding work / adding in the tasks, then
we're probably at the beginning.  So, we should weight the overall
estimate accordingly.  I still need to decide whether that should affect
the average or all.