Added the bones for the Westrum Report
We will want to have more than one report, so it seems useful to be able
to have a sub-route.  I have added the start of the second report, but
not the UI for it.

I also need to deal with the fact that the reports sub-route is not
included in the 'return to' part of the query param.  That would be
a nice touch, but I don't know if that's reasonable or possible...